Refactoring the concor contribution rest endpoint and merging this into the main controller by removing the test-data rest controller.
