+++ maat-court-data-api/src/main/resources/application.yaml
@@ -120,7 +120,7 @@ api-endpoints:
   application-domain: /api/internal/v1/application
   applicant-domain: /api/internal/v1/applicant
   debt-collection-enforcement-domain: /api/internal/v1/debt-collection-enforcement
-  debt-collection-enforcement-domain-test-data: /api/internal/v1/debt-collection-enforcement/test-data
+
 
 feature:
   postMvpEnabled: ${POST_MVP_ENABLED}
